整理一版本， 下个版本开始做个demo+测试。

+++ mtqueue.h
@@ -4,7 +4,7 @@
 
 struct request_message {
 	int type;
-	uint32_t source;
+	uintptr_t source;
 	int session;
 	void * data;
 	size_t sz;
